Extract account controller to its own crate
