+++ nym-vpn-core/crates/nym-account-controller/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "nym-account-controller"
+version.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+documentation.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+nym-config.workspace = true
+nym-http-api-client.workspace = true
+nym-vpn-api-client = { path = "../nym-vpn-api-client" }
+nym-vpn-store = { path = "../nym-vpn-store" }
+thiserror.workspace = true
+tokio-util.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+url.workspace = true
