+++ nym-vpn-core/crates/nym-account-controller/src/error.rs
@@ -1,0 +1,14 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+#[derive(Debug, thiserror::Error)]
+pub enum Error {
+    #[error("failed to get account summary")]
+    FailedToGetAccountSummary,
+
+    #[error("missing API URL")]
+    MissingApiUrl,
+
+    #[error("invalid API URL")]
+    InvalidApiUrl,
+}
