+++ nym-vpn-core/crates/nym-vpnd/src/main.rs
@@ -1,7 +1,7 @@
 // Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-mod account;
+// mod account;
 mod cli;
 mod command_interface;
 mod logging;
